moved vbcomponent extensions from reflection to extensions
