+++ RetailCoder.VBE/Extensions/VBComponentExtensions.cs
@@ -2,8 +2,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.InteropServices;
 using Microsoft.Vbe.Interop;
+using Rubberduck.Reflection;
 
-namespace Rubberduck.Reflection
+namespace Rubberduck.Extensions
 {
     [ComVisible(false)]
     public static class VBComponentExtensions
