+++ RetailCoder.VBE/UnitTesting/NewTestMethodCommand.cs
@@ -2,6 +2,7 @@ using System.Linq;
 using System.Runtime.InteropServices;
 using Microsoft.Vbe.Interop;
 using Rubberduck.Reflection;
+using Rubberduck.Extensions;
 
 namespace Rubberduck.UnitTesting
 {
